2022/March/pyproject.toml
[tool.poetry]
name = "march"
version = "0.1.0"
description = ""
authors = ["mauza <mauza@mauza.net>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
pandas = "^1.4.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
